pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "A template for building MCP servers with FastMCP"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
keywords = ["mcp", "model-context-protocol", "fastmcp", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.5",
    "python-dateutil>=2.9",
    "pytz>=2024.1",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "typing-extensions>=4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.3",
    "black>=24.0",
    "pre-commit>=3.6",
]

agent = [
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.2",
    "langchain[openai]>=0.1",
]

[project.scripts]
mcp-server = "mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "src/mcp_server",
    "README.md",
    "LICENSE",
    "pyproject.toml",
    ".env.example",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.3",
    "black>=24.0",
    "ipython>=8.20",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src/mcp_server",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "langchain.*",
    "langgraph.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ARG",
    "SIM",
]
ignore = [
    "E501",
    "B008",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\\.pyi?$'
